Refactors eeaGetTransactionReceipt to PrivGetTransactionReceipt
