+++ acceptance-tests/dsl/src/main/java/org/hyperledger/besu/tests/acceptance/dsl/privacy/transaction/PrivacyTransactions.java
@@ -13,15 +13,15 @@
 package org.hyperledger.besu.tests.acceptance.dsl.privacy.transaction;
 
 import org.hyperledger.besu.tests.acceptance.dsl.privacy.PrivacyNode;
-import org.hyperledger.besu.tests.acceptance.dsl.privacy.condition.EeaGetTransactionReceiptTransaction;
+import org.hyperledger.besu.tests.acceptance.dsl.privacy.condition.PrivGetTransactionReceiptTransaction;
 
 import java.util.List;
 
 public class PrivacyTransactions {
 
-  public EeaGetTransactionReceiptTransaction getPrivateTransactionReceipt(
+  public PrivGetTransactionReceiptTransaction getPrivateTransactionReceipt(
       final String transactionHash) {
-    return new EeaGetTransactionReceiptTransaction(transactionHash);
+    return new PrivGetTransactionReceiptTransaction(transactionHash);
   }
 
   public CreatePrivacyGroupTransaction createPrivacyGroup(
