+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/JsonRpcMethodsFactory.java
@@ -90,7 +90,7 @@ import org.hyperledger.besu.ethereum.api.jsonrpc.internal.methods.permissioning.
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.methods.permissioning.PermRemoveNodesFromWhitelist;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.parameters.JsonRpcParameter;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.privacy.methods.eea.EeaGetTransactionCount;
-import org.hyperledger.besu.ethereum.api.jsonrpc.internal.privacy.methods.eea.EeaGetTransactionReceipt;
+import org.hyperledger.besu.ethereum.api.jsonrpc.internal.privacy.methods.priv.PrevGetTransactionReceipt;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.privacy.methods.eea.EeaPrivateNonceProvider;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.privacy.methods.eea.EeaSendRawTransaction;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.privacy.methods.priv.PrivCreatePrivacyGroup;
@@ -363,7 +363,6 @@ public class JsonRpcMethodsFactory {
       if (eea) {
         addMethods(
             enabledMethods,
-            new EeaGetTransactionReceipt(blockchainQueries, enclave, parameter, privacyParameters),
             new EeaSendRawTransaction(privateTransactionHandler, transactionPool, parameter),
             new EeaGetTransactionCount(
                 parameter, new EeaPrivateNonceProvider(enclave, privateTransactionHandler)));
@@ -371,6 +370,7 @@ public class JsonRpcMethodsFactory {
       if (priv) {
         addMethods(
             enabledMethods,
+            new PrevGetTransactionReceipt(blockchainQueries, enclave, parameter, privacyParameters),
             new PrivCreatePrivacyGroup(
                 new Enclave(privacyParameters.getEnclaveUri()), privacyParameters, parameter),
             new PrivDeletePrivacyGroup(
