+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/privacy/methods/priv/PrevGetTransactionReceipt.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.hyperledger.besu.ethereum.api.jsonrpc.internal.privacy.methods.eea;
+package org.hyperledger.besu.ethereum.api.jsonrpc.internal.privacy.methods.priv;
 
 import static org.apache.logging.log4j.LogManager.getLogger;
 
@@ -48,7 +48,7 @@ import java.util.Optional;
 
 import org.apache.logging.log4j.Logger;
 
-public class EeaGetTransactionReceipt implements JsonRpcMethod {
+public class PrevGetTransactionReceipt implements JsonRpcMethod {
 
   private static final Logger LOG = getLogger();
 
@@ -57,7 +57,7 @@ public class EeaGetTransactionReceipt implements JsonRpcMethod {
   private final JsonRpcParameter parameters;
   private final PrivacyParameters privacyParameters;
 
-  public EeaGetTransactionReceipt(
+  public PrevGetTransactionReceipt(
       final BlockchainQueries blockchain,
       final Enclave enclave,
       final JsonRpcParameter parameters,
@@ -70,12 +70,12 @@ public class EeaGetTransactionReceipt implements JsonRpcMethod {
 
   @Override
   public String getName() {
-    return RpcMethod.EEA_GET_TRANSACTION_RECEIPT.getMethodName();
+    return RpcMethod.PRIV_GET_TRANSACTION_RECEIPT.getMethodName();
   }
 
   @Override
   public JsonRpcResponse response(final JsonRpcRequest request) {
-    LOG.trace("Executing {}", RpcMethod.EEA_GET_TRANSACTION_RECEIPT.getMethodName());
+    LOG.trace("Executing {}", RpcMethod.PRIV_GET_TRANSACTION_RECEIPT.getMethodName());
     final Hash transactionHash = parameters.required(request.getParams(), 0, Hash.class);
     final Optional<TransactionLocation> maybeLocation =
         blockchain.getBlockchain().getTransactionLocation(transactionHash);
